worker_000/requirements.txt
pyzmq
influxdb
blosc
pandas
numpy
scipy
spectrum